Add table support for capability "type" property (#12622)
* Add table type support

* Factor out yaml logic

* Rename workflow -> models
