+++ .changeset/new-cheetahs-sell.md
@@ -1,0 +1,5 @@
+---
+"chainlink": minor
+---
+
+Add table support to "type" property for step definitions
