+++ core/services/workflows/testdata/fixtures/workflows/marshalling/workflow_2.yaml
@@ -1,0 +1,28 @@
+ triggers:
+   - type: on_mercury_report@1
+     ref: report_data
+     config: {}
+
+ # no actions
+
+ consensus:
+   - type: 
+      name: trigger_test
+      version: "2"
+      tags:
+        chain: ethereum
+        aaShouldBeFirst: "true"
+        network: mainnet
+     config: {}
+     inputs:
+       observations:
+         - triggers.report_data.outputs
+
+ targets:
+   - type: write_polygon_mainnet@1
+     config: {}
+     inputs:
+       report:
+         - consensus.evm_median.outputs.report
+
+# yaml-language-server: $schema=../workflow_schema.json
